requirements.txt
Django
python-dotenv
pillow
django-crispy-forms
crispy-tailwind
google-cloud-dialogflow
python-dotenv
django-compressor
django-tailwind